+++ src/utils/FeatureFlags.ts
@@ -1,18 +1,20 @@
 import { NativeInstabug } from '../native/NativeInstabug';
-import { _registerW3CFlagsChangeListener } from '../modules/Instabug';
+import { _registerFeatureFlagsChangeListener } from '../modules/Instabug';
 
 export const FeatureFlags = {
   isW3ExternalTraceID: () => NativeInstabug.isW3ExternalTraceIDEnabled(),
   isW3ExternalGeneratedHeader: () => NativeInstabug.isW3ExternalGeneratedHeaderEnabled(),
   isW3CaughtHeader: () => NativeInstabug.isW3CaughtHeaderEnabled(),
+  networkLogLimit: () => NativeInstabug.getNetworkBodyMaxSize(),
 };
 
-export const registerW3CFlagsListener = () => {
-  _registerW3CFlagsChangeListener(
+export const registerFeatureFlagsListener = () => {
+  _registerFeatureFlagsChangeListener(
     (res: {
       isW3ExternalTraceIDEnabled: boolean;
       isW3ExternalGeneratedHeaderEnabled: boolean;
       isW3CaughtHeaderEnabled: boolean;
+      networkBodyLimit: number;
     }) => {
       FeatureFlags.isW3ExternalTraceID = async () => {
         return res.isW3ExternalTraceIDEnabled;
@@ -22,6 +24,9 @@ export const registerW3CFlagsListener = () => {
       };
       FeatureFlags.isW3CaughtHeader = async () => {
         return res.isW3CaughtHeaderEnabled;
+      };
+      FeatureFlags.networkLogLimit = async () => {
+        return res.networkBodyLimit;
       };
     },
   );